Make GitCommitChecker language agnostic
